rename responsive table to dataview
